requirements.txt
googlemaps
redis
PyJWT
celery
aliyunsdkcore
asn1crypto==0.24.0
astroid==2.1.0
atomicwrites==1.2.1
attrs==18.2.0
backcall==0.1.0
beautifulsoup4==4.6.3
bs4==0.0.1
certifi==2018.11.29
cffi==1.11.5
chardet==3.0.4
Click==7.0
colorama==0.4.1
cryptography==2.4.2
decorator==4.3.0
Django==2.2
django-cors-headers==2.4.0
elasticsearch==7.0.2
idna==2.8
ipykernel==5.1.0
ipython==7.1.1
ipython-genutils==0.2.0
isort==4.3.4
jedi==0.13.1
jupyter-client==5.2.3
jupyter-console==6.0.0
jupyter-core==4.4.0
lazy-object-proxy==1.3.1
logger==1.4
lxml==4.2.5
mccabe==0.6.1
more-itertools==4.3.0
neo4j-driver==1.6.2
neotime==1.0.0
numpy==1.16.0
pandas==0.23.4
parso==0.3.1
peewee==3.9.3
pickleshare==0.7.5
pluggy==0.8.0
prompt-toolkit==1.0.15
py==1.7.0
py2neo==4.1.3
pycparser==2.19
Pygments==2.3.0
pylint==2.2.2
PyMySQL==0.9.2
pytest==4.0.1
python-dateutil==2.7.5
pytrends==4.4.0
pytz==2018.7
pyzmq==17.1.2
requests==2.21.0
six==1.11.0
sqlparse==0.3.0
tornado==5.1.1
traitlets==4.3.2
typed-ast==1.1.1
urllib3==1.22
wcwidth==0.1.7
wrapt==1.11.0
psycopg2==2.8.2
pycryptodome==3.8.2